+++ core/language/en/gallery.php
@@ -123,6 +123,7 @@ $lang = array_merge($lang, [
 	'FILETYPES_PNG'              => 'png',
 	'FILETYPES_WEBP'             => 'webp',
 	'FILETYPES_AVIF'             => 'avif',
+	'FILETYPES_BMP'              => 'bmp',
 	'FILETYPES_ZIP'              => 'zip',
 
 	'FULL_EDITOR_GALLERY' => 'Full editor',