+++ core/language/nl/gallery.php
@@ -121,6 +121,7 @@ $lang = array_merge($lang, [
 	'FILETYPES_PNG'              => 'png',
 	'FILETYPES_WEBP'             => 'webp',
 	'FILETYPES_AVIF'             => 'avif',
+	'FILETYPES_BMP'              => 'bmp',
 	'FILETYPES_ZIP'              => 'zip',
 
 	'FULL_EDITOR_GALLERY' => 'Volledige bewerker',