+++ core/language/pt_preao/gallery.php
@@ -117,6 +117,7 @@ $lang = array_merge($lang, [
 	'FILETYPES_PNG'                => 'png',
 	'FILETYPES_WEBP'               => 'webp',
 	'FILETYPES_AVIF'               => 'avif',
+	'FILETYPES_BMP'                => 'bmp',
 	'FILETYPES_ZIP'                => 'zip',
 	'FULL_EDITOR_GALLERY'          => 'Editor completo',
 	'GALLERY_IMAGE'                => 'Imagem',